Move mock_user to tests
