+++ tests/test_sql.py
@@ -7,7 +7,7 @@ import os
 import random
 import unittest
 
-from mock_user import UserQuery
+from tests.mock_user import UserQuery
 from async_test import async_test
 from pypika import Query, Tables
